Implement CreateClientOptionsParser for Starknet
