+++ relayer/Cargo.toml
@@ -164,34 +164,34 @@ cgp-sync                = { git = "https://github.com/contextgeneric/cgp.git" }
 cgp-inner               = { git = "https://github.com/contextgeneric/cgp.git" }
 cgp-error-eyre          = { git = "https://github.com/contextgeneric/cgp.git" }
 
-hermes-chain-components              = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-chain-type-components         = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-relayer-components            = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-relayer-components-extra      = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-runtime-components            = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-async-runtime-components      = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-tokio-runtime-components      = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-runtime                       = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-error                         = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-encoding-components           = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-protobuf-encoding-components  = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-logging-components            = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-tracing-logging-components    = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-logger                        = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-test-components               = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-ibc-test-suite                = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-any-counterparty              = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cosmos-chain-components       = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cosmos-relayer                = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cosmos-wasm-relayer           = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cosmos-test-components        = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cosmos-encoding-components    = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cosmos-integration-tests      = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-comet-light-client-components = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-comet-light-client-context    = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-wasm-test-components          = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-wasm-client-components        = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-wasm-encoding-components      = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cli-components                = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cli                           = { git = "https://github.com/informalsystems/hermes-sdk.git" }
-hermes-cli-framework                 = { git = "https://github.com/informalsystems/hermes-sdk.git" }
+hermes-chain-components              = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-chain-type-components         = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-relayer-components            = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-relayer-components-extra      = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-runtime-components            = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-async-runtime-components      = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-tokio-runtime-components      = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-runtime                       = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-error                         = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-encoding-components           = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-protobuf-encoding-components  = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-logging-components            = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-tracing-logging-components    = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-logger                        = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-test-components               = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-ibc-test-suite                = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-any-counterparty              = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cosmos-chain-components       = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cosmos-relayer                = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cosmos-wasm-relayer           = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cosmos-test-components        = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cosmos-encoding-components    = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cosmos-integration-tests      = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-comet-light-client-components = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-comet-light-client-context    = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-wasm-test-components          = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-wasm-client-components        = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-wasm-encoding-components      = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cli-components                = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cli                           = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
+hermes-cli-framework                 = { git = "https://github.com/informalsystems/hermes-sdk.git", branch = "soares/cli-improvements-3" }
