+++ relayer/crates/starknet-cli/Cargo.toml
@@ -47,6 +47,7 @@ serde_json         = { workspace = true }
 clap               = { workspace = true, features = [ "derive" ] }
 eyre               = { workspace = true }
 toml               = { workspace = true }
+hex                = { workspace = true }
 tokio              = { workspace = true }
 tracing            = { workspace = true }
 tracing-subscriber = { workspace = true }
