+++ relayer/crates/starknet-cli/src/impls/mod.rs
@@ -1,3 +1,4 @@
 pub mod bootstrap;
 pub mod build;
+pub mod create_client;
 pub mod error;
