+++ relayer/scripts/create-client-b.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-cargo run --bin hermes -- \
+cargo run --bin hermes-starknet -- \
     --config test-data/config.toml \
-    client create \
-    --target-chain-id chain-b \
-    --counterparty-chain-id chain-a
+    create client \
+    --target-chain-id cosmos \
+    --counterparty-chain-id 393402133025997798000961
